Removing commented code, TODOs
